idek

+++ root.py
@@ -1,7 +1,9 @@
+import numpy as np
+
 class Root():
     def __init__(self, real, imag, color=(0,0,0,)):
         self.real = real
         self.imag = imag
-        self.complex = real + imag * 1j
+        self.complex = np.complex64(real + imag * 1j)
         self.color = color
         self.radius = 20
